+++ pcs_test/tier1/legacy/test_constraints.py
@@ -539,7 +539,7 @@ Ticket Constraints:
         )
 
         output, returnVal = pcs(
-            self.temp_cib.name, "constraint show --full".split()
+            self.temp_cib.name, "constraint config --full".split()
         )
         assert returnVal == 0
         ac(
@@ -679,7 +679,7 @@ Ticket Constraints:
         ), output
 
         output, returnVal = pcs(
-            self.temp_cib.name, "constraint location show --full".split()
+            self.temp_cib.name, "constraint location config --full".split()
         )
         ac(
             output,
@@ -705,7 +705,7 @@ Ticket Constraints:
         assert returnVal == 0
 
         output, returnVal = pcs(
-            self.temp_cib.name, "constraint location show --full".split()
+            self.temp_cib.name, "constraint location config --full".split()
         )
         ac(output, "Location Constraints:\n")
         assert returnVal == 0
@@ -2430,7 +2430,7 @@ Ticket Constraints:
 
         output, returnVal = pcs(
             self.temp_cib.name,
-            "constraint location show resources dummy --full".split(),
+            "constraint location config resources dummy --full".split(),
         )
         ac(
             output,
@@ -2491,7 +2491,7 @@ Ticket Constraints:
 
         output, returnVal = pcs(
             self.temp_cib.name,
-            "constraint location show resources dummy --full".split(),
+            "constraint location config resources dummy --full".split(),
         )
         ac(
             output,
@@ -3694,7 +3694,7 @@ class TicketCreateWithSet(ConstraintBaseTest):
             "constraint ticket set A B setoptions ticket=T".split()
         )
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=[
                 "Ticket Constraints:",
                 "  Resource Sets:",
@@ -3730,7 +3730,7 @@ class TicketAdd(ConstraintBaseTest):
             "constraint ticket add T master A loss-policy=fence".split()
         )
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=[
                 "Ticket Constraints:",
                 "  Master A loss-policy=fence ticket=T",
@@ -3778,7 +3778,7 @@ class TicketAdd(ConstraintBaseTest):
             ],
         )
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=[
                 "Ticket Constraints:",
                 "  Master A loss-policy=fence ticket=T",
@@ -3818,7 +3818,7 @@ class TicketDeleteRemoveTest(ConstraintBaseTest):
             "constraint ticket set A setoptions ticket=T".split()
         )
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=[
                 "Ticket Constraints:",
                 "  A ticket=T",
@@ -3835,7 +3835,7 @@ class TicketDeleteRemoveTest(ConstraintBaseTest):
         )
 
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=[
                 "Ticket Constraints:",
                 "  Resource Sets:",
@@ -3845,7 +3845,7 @@ class TicketDeleteRemoveTest(ConstraintBaseTest):
 
     def _test_fail_when_no_matching_ticket_constraint_here(self):
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             stdout_full=["Ticket Constraints:"],
         )
         self.assert_pcs_fail(
@@ -3875,7 +3875,7 @@ class TicketShow(ConstraintBaseTest):
             "constraint ticket add T master A loss-policy=fence".split()
         )
         self.assert_pcs_success(
-            "constraint ticket show".split(),
+            "constraint ticket config".split(),
             [
                 "Ticket Constraints:",
                 "  Master A loss-policy=fence ticket=T",
@@ -4086,7 +4086,7 @@ class LocationShowWithPattern(ConstraintBaseTest):
     def test_show(self):
         self.fixture()
         self.assert_pcs_success(
-            "constraint location show --all --full".split(),
+            "constraint location config --all --full".split(),
             outdent(
                 """\
             Location Constraints:
@@ -4138,7 +4138,7 @@ class LocationShowWithPattern(ConstraintBaseTest):
         )
 
         self.assert_pcs_success(
-            "constraint location show".split(),
+            "constraint location config".split(),
             outdent(
                 """\
             Location Constraints:
@@ -4190,7 +4190,7 @@ class LocationShowWithPattern(ConstraintBaseTest):
         )
 
         self.assert_pcs_success(
-            "constraint location show nodes --full".split(),
+            "constraint location config nodes --full".split(),
             outdent(
                 # pylint:disable=trailing-whitespace
                 """\
@@ -4250,7 +4250,7 @@ class LocationShowWithPattern(ConstraintBaseTest):
         )
 
         self.assert_pcs_success(
-            "constraint location show nodes node2".split(),
+            "constraint location config nodes node2".split(),
             outdent(
                 """\
             Location Constraints:
@@ -4283,7 +4283,7 @@ class LocationShowWithPattern(ConstraintBaseTest):
         )
 
         self.assert_pcs_success(
-            "constraint location show resources regexp%R_[0-9]+".split(),
+            "constraint location config resources regexp%R_[0-9]+".split(),
             outdent(
                 """\
             Location Constraints: